requirements.txt
streamlit>=1.33
pydantic>=2.6
pandas>=2.2
numpy>=1.26
Pillow>=10.0
torch>=2.0
torchvision>=0.15
